add reset button; extract game board
